requirements.txt
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
requests==2.28.2
peewee==3.16.0
matplotlib==3.7.1